% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginate_listing.R
\name{paginate_listing}
\alias{paginate_listing}
\title{Paginate listings}
\usage{
paginate_listing(
  lsting,
  lpp = 15,
  cpp = NULL,
  min_siblings = 2,
  nosplitin = character(),
  colwidths = NULL,
  verbose = FALSE
)
}
\arguments{
\item{lsting}{(\code{listing_df})\cr the listing to paginate.}

\item{lpp}{(\code{number} or \code{NULL})\cr number of row lines (not counting titles and
footers) to have per page. Standard is \code{15} while \code{NULL} disables vertical
pagination.}

\item{cpp}{numeric(1). Number of columns per page}

\item{min_siblings}{numeric. Minimum sibling rows which must appear on either side of pagination row for a
mid-subtable split to be valid. Defaults to 2.}

\item{nosplitin}{character. List of names of sub-tables where page-breaks are not allowed, regardless of other
considerations. Defaults to none.}

\item{colwidths}{(\code{numeric} or \code{NULL})\cr print widths of columns, if manually
set/previously known.}

\item{verbose}{logical(1). Should additional informative messages about the search for
pagination breaks be shown. Defaults to \code{FALSE}.}
}
\value{
A list of listings' objects that are meant to be on separated pages.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Pagination of a listing. This can be vertical for long listings with many
rows or horizontal if there are many columns.
}
\examples{
# Create a standard listing
library(tibble)
dat <- ex_adae

lsting <- as_listing(dat[1:25, ],
  key_cols = c("USUBJID", "AGE", "AESOC")
) \%>\%
  add_listing_col("AETOXGR") \%>\%
  add_listing_col("BMRKR1", format = "xx.x") \%>\%
  add_listing_col("AESER / AREL", fun = function(df) paste(df$AESER, df$AREL, sep = " / "))

# Vertical pagination
paginate_listing(lsting, lpp = 10)

# Horizontal pagination
# paginate_listing(lsting, cpp = 10, lpp = 40)

}
