% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlistings.R
\name{split_listing_by_param}
\alias{split_listing_by_param}
\title{Split Listing into a List of Listings by Values of a Parameter}
\usage{
split_listing_by_param(lsting, param, page_prefix = param)
}
\arguments{
\item{lsting}{listing_df. The listing to split.}

\item{param}{character. Name of the parameter upon which to split.}

\item{page_prefix}{character. Prefix to be appended with the split value (\code{param} level),
at the end of the subtitles, corresponding to each resulting list element (listing).}
}
\value{
A list of \code{lsting_df} objects each corresponding to a unique value of \code{param}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Split is performed based on unique values of the given parameter present in the listing.
}
\examples{
dat <- ex_adae[1:20, ]

lsting <- as_listing(
    dat,
    key_cols = c("USUBJID", "AGE"),
    disp_cols = "SEX",
    main_title = "title",
    main_footer = "footer"
  ) \%>\%
  add_listing_col("BMRKR1", format = "xx.x") \%>\%
  split_listing_by_param("SEX")

lsting

}
