% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_as_docx.R
\name{to_flextable.data.frame}
\alias{to_flextable.data.frame}
\title{data.frame method to to_flextable}
\usage{
\method{to_flextable}{data.frame}(x, col_width = NULL, font_size = 9, ...)
}
\arguments{
\item{x}{object to to_flextable}
}
\description{
data.frame method to to_flextable
}
