% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_as_docx.R
\name{to_flextable.default}
\alias{to_flextable.default}
\title{default method to to_flextable}
\usage{
\method{to_flextable}{default}(x, ...)
}
\arguments{
\item{x}{object to to_flextable}

\item{...}{additional arguments. not used.}
}
\description{
default method to to_flextable
}
