% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_as_docx.R
\name{to_flextable.listing_df}
\alias{to_flextable.listing_df}
\title{convert listing to flextable}
\usage{
\method{to_flextable}{listing_df}(x, cpp = 100, lpp = 100, ...)
}
\description{
convert listing to flextable
}
