% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_as_docx.R
\name{to_flextable}
\alias{to_flextable}
\title{s3 method for to_flextable}
\usage{
to_flextable(x, ...)
}
\arguments{
\item{x}{object to to_flextable}

\item{...}{additional arguments passed to methods}
}
\description{
s3 method for to_flextable
}
