% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlistings_methods.R
\name{toString,listing_df-method}
\alias{toString,listing_df-method}
\title{toString method for listing_df object}
\usage{
\S4method{toString}{listing_df}(x, widths = NULL, col_gap = 3, hsep = default_hsep())
}
\arguments{
\item{x}{listing_df object}

\item{widths}{numeric (or  NULL). (proposed) widths for the columns
of \code{x}. The expected length  of this numeric vector can be
retrieved with  \code{ncol() + 1}  as the  column of row  names must
also be considered.}

\item{col_gap}{numeric(1). Space (in characters) between columns}

\item{hsep}{character(1).   Characters  to  repeat   to  create
header/body separator line.}
}
\value{
A character value containing the listing rendered into
ASCII text.
}
\description{
toString method for listing_df object
}
\examples{
lsting <- as_listing(mtcars)
toString(lsting)
}
