% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlistings_methods.R
\name{format_colvector}
\alias{format_colvector}
\alias{vec_nlines}
\alias{vec_nlines,ANY-method}
\title{Utilities for formatting a listing column}
\usage{
format_colvector(df, colnm, colvec = df[[colnm]])

vec_nlines(vec, max_width = NULL)

\S4method{vec_nlines}{ANY}(vec, max_width = NULL)
}
\arguments{
\item{df}{listing_df. The listing.}

\item{colnm}{Column name}

\item{colvec}{Column values based on colnm}

\item{vec}{A vector.}

\item{max_width}{numeric (or NULL). The width the column will be
rendered in.}
}
\value{
a numeric vector of the number of lines elementwise that
will be needed to render the elements of \code{vec} to width
\code{max_width}.
}
\description{
For \code{vec_nlines}, calculate the number of lines each element of a column vector will
take to render. For \code{format_colvector},
}
\keyword{internal}
