% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginate_listing.R
\name{pag_listing_indices}
\alias{pag_listing_indices}
\title{Defunct functions}
\usage{
pag_listing_indices(
  lsting,
  lpp = 15,
  colwidths = NULL,
  max_width = NULL,
  verbose = FALSE
)
}
\arguments{
\item{lsting}{listing_df. The listing to paginate.}

\item{lpp}{numeric(1) or NULL. Number of row lines (not counting titles and
footers) to have per page. Standard is \code{70} while \code{NULL} disables vertical
pagination.}

\item{colwidths}{numeric vector.  Column widths (in characters) for
use with vertical pagination.}

\item{max_width}{integer(1), character(1) or NULL. Width that title
and   footer   (including   footnotes)  materials   should   be
word-wrapped to. If NULL, it is  set to the current print width
of the  session (\code{getOption("width")}). If set to \code{"auto"},
the width of the table (plus any table inset) is used. Ignored
completely if \code{tf_wrap} is \code{FALSE}.}

\item{verbose}{logical(1). Should additional informative messages about the search for
pagination breaks be shown. Defaults to \code{FALSE}.}
}
\description{
These functions are defunct and their symbols will be removed entirely
in a future release.
}
