% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlistings.R
\name{listing_methods}
\alias{listing_methods}
\alias{print.listing_df}
\alias{make_row_df,listing_df-method}
\alias{[,listing_df-method}
\alias{main_title,listing_df-method}
\alias{subtitles,listing_df-method}
\alias{main_footer,listing_df-method}
\alias{prov_footer,listing_df-method}
\alias{paginate_listing}
\title{Methods for `listing_df` objects}
\usage{
\method{print}{listing_df}(x, ...)

\S4method{make_row_df}{listing_df}(
  tt,
  colwidths = NULL,
  visible_only = TRUE,
  rownum = 0,
  indent = 0L,
  path = character(),
  incontent = FALSE,
  repr_ext = 0L,
  repr_inds = integer(),
  sibpos = NA_integer_,
  nsibs = NA_integer_
)

\S4method{[}{listing_df}(x, i, j, drop = FALSE)

\S4method{main_title}{listing_df}(obj)

\S4method{subtitles}{listing_df}(obj)

\S4method{main_footer}{listing_df}(obj)

\S4method{prov_footer}{listing_df}(obj)

paginate_listing(
  lsting,
  lpp = 15,
  min_siblings = 2,
  nosplitin = character(),
  colwidths = NULL,
  verbose = FALSE
)
}
\arguments{
\item{x}{listing_df. The listing.}

\item{...}{dots. Unused}

\item{tt}{ANY. Object representing the table-like object to be summarized.}

\item{colwidths}{numeric. Print  widths of  columns, if  manually
set/previously known.}

\item{visible_only}{logical(1). Should only visible aspects of the table structure be reflected in this summary. Defaults to \code{TRUE}. May not be supported by all methods.}

\item{rownum}{numeric(1). Internal detail do not set manually.}

\item{indent}{integer(1). Internal detail do not set manually.}

\item{path}{character.  Path  to  the (sub)table  represented  by
\code{tt}. Defaults to \code{character()}}

\item{incontent}{logical(1). Internal detail do not set manually.}

\item{repr_ext}{integer(1). Internal detail do not set manually.}

\item{repr_inds}{integer. Internal detail do not set manually.}

\item{sibpos}{integer(1). Internal detail do not set manually.}

\item{nsibs}{integer(1). Internal detail do not set manually.}

\item{i}{ANY. Passed to base `[` methods.}

\item{j}{ANY. Passed to base `[` methods.}

\item{drop}{For matrices and arrays.  If \code{TRUE} the result is
    coerced to the lowest possible dimension (see the examples).  This
    only works for extracting elements, not for the replacement.  See
    \code{\link[base]{drop}} for further details.
  }

\item{obj}{The object.}

\item{lsting}{listing_df. The listing to paginate.}

\item{lpp}{numeric(1). Number of row lines (not counting titles and
footers) to have per page.}

\item{min_siblings}{numeric. Minimum sibling rows which must appear on either side of pagination row for a
mid-subtable split to be valid. Defaults to 2.}

\item{nosplitin}{character. List of names of sub-tables where page-breaks are not allowed, regardless of other
considerations. Defaults to none.}

\item{verbose}{logical(1). Should additional informative messages about the search for
pagination breaks be shown. Defaults to \code{FALSE}.}
}
\description{
See core documentation in \code{formatters} for descriptions
of these functions.
}
