% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlistings_methods.R
\name{listing_methods}
\alias{listing_methods}
\alias{print.listing_df}
\alias{format_colvector}
\alias{vec_nlines}
\alias{vec_nlines,ANY-method}
\alias{make_row_df,listing_df-method}
\alias{[,listing_df-method}
\alias{main_title,listing_df-method}
\alias{subtitles,listing_df-method}
\alias{main_footer,listing_df-method}
\alias{prov_footer,listing_df-method}
\alias{main_title<-,listing_df-method}
\alias{subtitles<-,listing_df-method}
\alias{main_footer<-,listing_df-method}
\alias{prov_footer<-,listing_df-method}
\title{Methods for \code{listing_df} objects}
\usage{
\method{print}{listing_df}(x, ...)

format_colvector(df, colnm, colvec = df[[colnm]])

vec_nlines(vec, max_width = NULL)

\S4method{vec_nlines}{ANY}(vec, max_width = NULL)

\S4method{make_row_df}{listing_df}(
  tt,
  colwidths = NULL,
  visible_only = TRUE,
  rownum = 0,
  indent = 0L,
  path = character(),
  incontent = FALSE,
  repr_ext = 0L,
  repr_inds = integer(),
  sibpos = NA_integer_,
  nsibs = NA_integer_
)

\S4method{[}{listing_df}(x, i, j, drop = FALSE)

\S4method{main_title}{listing_df}(obj)

\S4method{subtitles}{listing_df}(obj)

\S4method{main_footer}{listing_df}(obj)

\S4method{prov_footer}{listing_df}(obj)

\S4method{main_title}{listing_df}(obj) <- value

\S4method{subtitles}{listing_df}(obj) <- value

\S4method{main_footer}{listing_df}(obj) <- value

\S4method{prov_footer}{listing_df}(obj) <- value
}
\arguments{
\item{x}{listing_df. The listing.}

\item{...}{dots. Unused}

\item{df}{listing_df. The listing.}

\item{colnm}{Column name}

\item{colvec}{Column values based on colnm}

\item{vec}{A vector.}

\item{max_width}{numeric(1) or NULL. Maximum width for title/footer
materials.}

\item{tt}{ANY. Object representing the table-like object to be summarized.}

\item{colwidths}{numeric. Internal detail do not set manually.}

\item{visible_only}{logical(1). Should only visible aspects of the table structure be reflected in this summary.
Defaults to \code{TRUE}. May not be supported by all methods.}

\item{rownum}{numeric(1). Internal detail do not set manually.}

\item{indent}{integer(1). Internal detail do not set manually.}

\item{path}{character.  Path  to  the (sub)table  represented  by
\code{tt}. Defaults to \code{character()}}

\item{incontent}{logical(1). Internal detail do not set manually.}

\item{repr_ext}{integer(1). Internal detail do not set manually.}

\item{repr_inds}{integer. Internal detail do not set manually.}

\item{sibpos}{integer(1). Internal detail do not set manually.}

\item{nsibs}{integer(1). Internal detail do not set manually.}

\item{i}{ANY. Passed to base \code{[} methods.}

\item{j}{ANY. Passed to base \code{[} methods.}

\item{drop}{For matrices and arrays.  If \code{TRUE} the result is
    coerced to the lowest possible dimension (see the examples).  This
    only works for extracting elements, not for the replacement.  See
    \code{\link[base]{drop}} for further details.
  }

\item{obj}{The object.}

\item{value}{typically an array-like \R object of a similar class as
    \code{x}.}
}
\description{
See core documentation in \code{formatters} for descriptions
of these functions.
}
