% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlistings_methods.R
\name{listing_methods}
\alias{listing_methods}
\alias{print.listing_df}
\alias{[,listing_df-method}
\alias{main_title,listing_df-method}
\alias{subtitles,listing_df-method}
\alias{main_footer,listing_df-method}
\alias{prov_footer,listing_df-method}
\alias{main_title<-,listing_df-method}
\alias{subtitles<-,listing_df-method}
\alias{main_footer<-,listing_df-method}
\alias{prov_footer<-,listing_df-method}
\title{Methods for \code{listing_df} objects}
\usage{
\method{print}{listing_df}(x, ...)

\S4method{[}{listing_df}(x, i, j, drop = FALSE)

\S4method{main_title}{listing_df}(obj)

\S4method{subtitles}{listing_df}(obj)

\S4method{main_footer}{listing_df}(obj)

\S4method{prov_footer}{listing_df}(obj)

\S4method{main_title}{listing_df}(obj) <- value

\S4method{subtitles}{listing_df}(obj) <- value

\S4method{main_footer}{listing_df}(obj) <- value

\S4method{prov_footer}{listing_df}(obj) <- value
}
\arguments{
\item{x}{listing_df. The listing.}

\item{...}{dots. Unused}

\item{i}{ANY. Passed to base \code{[} methods.}

\item{j}{ANY. Passed to base \code{[} methods.}

\item{drop}{For matrices and arrays.  If \code{TRUE} the result is
    coerced to the lowest possible dimension (see the examples).  This
    only works for extracting elements, not for the replacement.  See
    \code{\link[base]{drop}} for further details.
  }

\item{obj}{The object.}

\item{value}{typically an array-like \R object of a similar class as
    \code{x}.}
}
\value{
for getter methods, the value of the aspect of
\code{obj}; for setter methods, \code{obj} with
the relevant element of the listing updated.
}
\description{
See core documentation in \code{formatters} for descriptions
of these functions.
}
\examples{

lsting <- as_listing(mtcars)
main_title(lsting) <- "Hi there"

main_title(lsting)
}
\keyword{internal}
