% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlistings.R
\name{as_listing}
\alias{as_listing}
\alias{as_keycol}
\alias{is_keycol}
\alias{get_keycols}
\alias{matrix_form,listing_df-method}
\alias{listing_dispcols}
\alias{add_listing_dispcol}
\alias{listing_dispcols<-}
\alias{add_listing_col}
\title{Create a Listing from a \code{data.frame} or \code{tibble}}
\usage{
as_listing(
  df,
  cols = key_cols,
  key_cols = names(df)[1],
  main_title = NULL,
  subtitles = NULL,
  main_footer = NULL,
  prov_footer = NULL
)

as_keycol(vec)

is_keycol(vec)

get_keycols(df)

\S4method{matrix_form}{listing_df}(obj, indent_rownames = FALSE)

listing_dispcols(df)

add_listing_dispcol(df, new)

listing_dispcols(df) <- value

add_listing_col(df, name, fun = NULL, format = NULL)
}
\arguments{
\item{df}{listing_df. The listing to modify.}

\item{cols}{(\code{character}) names of columns (including but not limited to key columns)
which should be displayed when the listing is rendered.}

\item{key_cols}{(\code{character}) names of columns which should be treated as \emph{key columns}
when rendering the listing.}

\item{main_title}{(\code{string} or \code{NULL}) the main title for the listing.}

\item{subtitles}{(\code{character} or \code{NULL}) a vector of subtitle(s) for the
listing.}

\item{main_footer}{(\code{character} or \code{NULL}) a vector of main footer lines
for the listing.}

\item{prov_footer}{(\code{character} or \code{NULL}) a vector of provenance strings
for the listing.}

\item{vec}{any. A column vector from a \code{listing_df}}

\item{obj}{ANY. Object to be transformed into a ready-to-render form (a MatrixPrintForm object)}

\item{indent_rownames}{logical(1), if TRUE the column with the row names in the `strings` matrix of has indented row
names (strings pre-fixed)}

\item{new}{character. Names of columns to be added to
the set of display columns.}

\item{value}{character. New value.}

\item{name}{character(1). Name of the existing or new column to be
displayed when the listing is rendered}

\item{fun}{function or NULL. A function which accepts \code{df} and
returns the vector for a new column, which is added to \code{df} as
\code{name}, or NULL if marking an existing column as
a listing column}

\item{format}{FormatSpec. A format specification (format string,
function, or \code{sprintf} format) for use when displaying the column
during rendering.}
}
\value{
A \code{listing_df} object, sorted by the key columns.

\code{df}, with \code{name} created (if necessary) and marked for
display during rendering.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Creates listings by using \code{cols} and \code{key_cols} to produce a compact and
elegant representation of the \code{data.frame} or \code{tibble} in input.
}
\examples{
dat <- ex_adae

lsting <- as_listing(dat[1:25, ], key_cols = c("USUBJID", "AESOC")) \%>\%
  add_listing_col("AETOXGR") \%>\%
  add_listing_col("BMRKR1", format = "xx.x") \%>\%
  add_listing_col("AESER / AREL", fun = function(df) paste(df$AESER, df$AREL, sep = " / "))

mat <- matrix_form(lsting)

cat(toString(mat))

}
