% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlistings-package.R
\docType{package}
\name{rlistings-package}
\alias{rlistings}
\alias{rlistings-package}
\title{rlistings: Clinical Trial Style Data Readout Listings}
\description{
Listings are often part of the submission of clinical trial data in regulatory settings. We provide a framework for the specific formatting features often used when displaying large datasets in that context.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/insightsengineering/rlistings}
  \item Report bugs at \url{http://github.com/insightsengineering/rlistings/issues}
}

}
\author{
\strong{Maintainer}: Gabriel Becker \email{gabembecker@gmail.com}

Authors:
\itemize{
  \item Adrian Waddell \email{adrian.waddell@gene.com}
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
