% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlistings-package.R
\docType{package}
\name{rlistings-package}
\alias{rlistings}
\alias{rlistings-package}
\title{rlistings: Clinical Trial Style Data Readout Listings}
\description{
Listings are often part of the submission of clinical trial data in regulatory settings. We provide a framework for the specific formatting features often used when displaying large datasets in that context.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://insightsengineering.github.io/rlistings/}
  \item \url{https://github.com/insightsengineering/rlistings/}
  \item Report bugs at \url{https://github.com/insightsengineering/rlistings/issues}
}

}
\author{
\strong{Maintainer}: Joe Zhu \email{joe.zhu@roche.com} (\href{https://orcid.org/0000-0001-7566-2787}{ORCID})

Authors:
\itemize{
  \item Gabriel Becker \email{gabembecker@gmail.com} (original creator of the package)
  \item Adrian Waddell \email{adrian.waddell@gene.com}
  \item Davide Garolini \email{davide.garolini@roche.com} (\href{https://orcid.org/0000-0002-1445-1369}{ORCID})
  \item Emily de la Rua \email{emily.de_la_rua@contractors.roche.com} (\href{https://orcid.org/0009-0000-8738-5561}{ORCID})
}

Other contributors:
\itemize{
  \item Abinaya Yogasekaram \email{abinaya.yogasekaram@contractors.roche.com} (\href{https://orcid.org/0009-0005-2083-1105}{ORCID}) [contributor]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
